[Engine] Values iterator should be 'double' and not 'int'

+++ core/engine.hpp
@@ -191,7 +191,7 @@ class engine {
   static vector<double> get_pre_neuron_values(int neuron_id, string variable) {
     vector<double> values;
     if(!pre_synaptic_lists.empty()) {
-      for(vector<int>::size_type index = 0; index < pre_synaptic_lists[neuron_id].size(); ++index) {
+      for(vector<double>::size_type index = 0; index < pre_synaptic_lists[neuron_id].size(); ++index) {
         values.push_back(synapse_value(pre_synaptic_lists[neuron_id][index], variable));
       }
     }
